requirements.txt
el-internationalisation @ git+https://github.com/enabling-languages/el_internationalisation.git@4cc23c143f862400d54ca65f6c1faad802a2297c
isodate==0.6.1
lxml==4.9.3
prettytable==3.8.0
PyICU==2.11
pymarc==5.1.0
pyparsing==3.1.0
rdflib==6.3.2
regex==2023.6.3
six==1.16.0
unicodedataplus==15.0.0.post2
wcwidth==0.2.6
